direct method

+++ 28. vslam_course/HW6/code/test_common.cpp
@@ -1,8 +1,8 @@
 /*
  * @Author: Liu Weilong
  * @Date: 2021-02-27 22:49:48
- * @LastEditors: Liu Weilong
- * @LastEditTime: 2021-02-27 23:54:55
+ * @LastEditors: Liu Weilong 
+ * @LastEditTime: 2021-03-03 09:30:01
  * @Description: 
  */
 
@@ -10,10 +10,10 @@
 
 int main()
 {
-    std::vector<Eigen::Vector2d,Eigen::aligned_allocator<Eigen::Vector2d>> temp;
+    std::vector<Eigen::Vector2f,Eigen::aligned_allocator<Eigen::Vector2f>> temp;
     std::vector<cv::Point2f> t;
     for(int i=0;i<10;i++)
-        temp.push_back(Eigen::Vector2d(i,i));
+        temp.push_back(Eigen::Vector2f(i,i));
     eigen2cv(temp,t);
     cout<<"in the opencv array"<<endl;
     for(auto & p:t)
